pyproject.toml
[project]
name = "api-banap-app"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "Asher Novelli", email = "asherndebortoli@gmail.com"},
]
dependencies = [
    "Flask>=3.0.3",
    "pillow>=10.3.0",
    "numpy>=1.26.4",
    "matplotlib>=3.9.0",
    "psycopg2>=2.9.10",
    "python-dotenv>=1.0.1",
]
requires-python = ">=3.12"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[tool.pdm]
distribution = true
